2xLED and 2xButton
Using breadboard wiring and a rapspi, the program looks for a
button press to determine what to do. The first button turns on the LED while
it is pressed. The other button (if held down) breaks the loop and,
in effect, ends the program. If you press both at the same time, a yellow LED
lights up. If you press the first btn and then the second, both light up.