LED Dice roller. 7xLED 2xButtons.
Used: 7xLED, 2xButtons, 10x M/F wires, 5x M/M wires, 7x 220 Ohm resistors, breadboard, rpi 3.
Program waits for user to press a button. The roll button triggers
the rollDice() method that rolls a random number 1-6 and outputs the dice roll
to the screen and through LEDs. The other button shutsdown the program and
triggers the GPIO cleanup. This can be used as a class in a later program.